chore(examples/npm-workspaces): use server-side type alias
This is to showcase a npm workspaces setup with working type aliases.
Type aliases within child packages are discouraged as they won't work
when tsc is initiated by some other child package.
