+++ examples/npm-workspaces/client/tsconfig.json
@@ -1,15 +1,9 @@
 {
+  "extends": "../tsconfig.json",
+  
+  "include": ["src"],
+  
   "compilerOptions": {
-    "target": "esnext",
-    "module": "NodeNext",
-    "moduleResolution": "NodeNext",
-    "esModuleInterop": true,
-    "strict": true,
     "outDir": "dist",
-    "skipLibCheck": true,
-    "paths": {
-      "@/*": ["./src/*"]
-    }
   },
-  "include": ["src"]
 }
