+++ examples/npm-workspaces/server/package.json
@@ -2,6 +2,7 @@
   "name": "@examples-npm-workspaces/server",
   "private": true,
   "type": "module",
+  "types": "src/index.ts",
   "scripts": {
     "build": "tsc",
     "dev": "tsx watch src/index.ts",
