+++ examples/npm-workspaces/server/src/_skel/db.ts
