+++ examples/npm-workspaces/server/src/index.ts
@@ -1,27 +1,12 @@
 /**
  * This a minimal tRPC server
  */
+ 
 import { createHTTPServer } from '@trpc/server/adapters/standalone';
-import { publicProcedure, router } from './trpc.js';
-import userRouter from './user.router.js';
+import appRouter from './app-router.ts';
 
-const appRouter = router({
-  user: userRouter,
-  examples: {
-    iterable: publicProcedure.query(async function* () {
-      for (let i = 0; i < 3; i++) {
-        await new Promise((resolve) => setTimeout(resolve, 500));
-        yield i;
-      }
-    }),
-  },
-});
+export type { AppRouter } from "./app-router.ts";
 
-// Export type router type signature, this is used by the client.
-export type AppRouter = typeof appRouter;
-
-const server = createHTTPServer({
-  router: appRouter,
-});
+const server = createHTTPServer({ router: appRouter });
 
 server.listen(3000);
